First draft of sprite run with arguments
Instead of running onClick as a label name, allow arguments to also be passed into it. This allows for more sprite functionality flexibility.
In my case, an intercept function that allows me to play a noise when something is clicked, then jump to a supplied label.

I haven't tested it because uuuuh i dont know how.